+++ UXToolsGame/Plugins/UXTools/Source/UXTools/Private/Behaviors/UxtTapToPlaceComponent.cpp
@@ -8,6 +8,7 @@
 #include "Engine/World.h"
 #include "Input/UxtFarPointerComponent.h"
 #include "Input/UxtInputSubsystem.h"
+#include "Kismet/KismetSystemLibrary.h"
 #include "Math/UnrealMathUtility.h"
 #include "Utils/UxtFunctionLibrary.h"
 
@@ -31,6 +32,15 @@ namespace
 		Axes.SetAxes(&Forward, &Right, &Up);
 		return FQuat(Axes);
 	}
+
+	float GetDefaultSurfaceNormalOffset(const USceneComponent* TargetComponent)
+	{
+		FVector BoundsCentre, BoxExtent;
+		float SphereRadius;
+		UKismetSystemLibrary::GetComponentBounds(TargetComponent, BoundsCentre, BoxExtent, SphereRadius);
+		float PivotPointOffset = TargetComponent->GetComponentLocation().X - BoundsCentre.X;
+		return BoxExtent.X - PivotPointOffset;
+	}
 } // namespace
 
 UUxtTapToPlaceComponent::UUxtTapToPlaceComponent()
@@ -50,8 +60,7 @@ void UUxtTapToPlaceComponent::SetTargetComponent(UPrimitiveComponent* Target)
 	TargetComponent.OverrideComponent = Target;
 	if (Target)
 	{
-		FBoxSphereBounds Bounds = Target->CalcLocalBounds();
-		SurfaceNormalOffset = Bounds.BoxExtent.X * Target->GetComponentTransform().GetScale3D().X;
+		DefaultSurfaceNormalOffset = GetDefaultSurfaceNormalOffset(Target);
 	}
 }
 
@@ -92,8 +101,7 @@ void UUxtTapToPlaceComponent::BeginPlay()
 
 	if (UPrimitiveComponent* Target = GetTargetComponent())
 	{
-		FBoxSphereBounds Bounds = Target->CalcLocalBounds();
-		SurfaceNormalOffset = Bounds.BoxExtent.X * Target->GetComponentTransform().GetScale3D().X;
+		DefaultSurfaceNormalOffset = GetDefaultSurfaceNormalOffset(Target);
 	}
 }
 
@@ -140,7 +148,8 @@ void UUxtTapToPlaceComponent::TickComponent(float DeltaTime, ELevelTick TickType
 			if (Result.GetComponent())
 			{
 				// Add SurfaceNormalOffset so object is placed touching the surface, rather than overlapping with it
-				HitPosition = Result.Location + Result.Normal * SurfaceNormalOffset;
+				HitPosition =
+					Result.Location + Result.Normal * (bUseDefaultSurfaceNormalOffset ? DefaultSurfaceNormalOffset : SurfaceNormalOffset);
 				const FVector TowardsOriginPose = Facing = OriginPose.GetLocation() - HitPosition;
 
 				// Check if the target surface is flat, e.g. floor or ceiling